removed old events for reactions and refactored !group Toggle
